selftests/bpf: Add regex helpers for selftests
Factor out a set of helper functions for regex matching to make them
available across BPF selftests.

match_regex does not pass any cflags to regcomp, so it does not use
extended syntax. The reasoning for this is that a lot of patterns in
tests match parentheses, and extended syntax requires escaping them.

To use the extended syntax, compile_regex helper may be used directly.